feature file nominal pitch in midi, changes to DP algorithm

+++ Resynthesis.h
@@ -21,7 +21,7 @@ typedef struct {
     float hz;
     float amp;
     int frame;
-    float nominal;
+    int nominal;
     //add other features?
 } AUDIO_FEATURE;
